personalización swagger UI

+++ LibreriaElPortal-WebAPI/Controllers/ClienteController.cs
@@ -13,14 +13,24 @@ namespace LibreriaElPortal_WebAPI.Controllers
     public class ClienteController : ControllerBase
     {
         public readonly IClienteRepository _clienteRepository;
+        public readonly IVentaRepository _ventaRepository;
         public readonly IMapper _mapper;
-        public ClienteController(IClienteRepository clienteRepository, IMapper mapper) 
+        public ClienteController(IClienteRepository clienteRepository, IMapper mapper, IVentaRepository ventaRepository) 
         {
             _clienteRepository = clienteRepository;
             _mapper = mapper;
+            _ventaRepository = ventaRepository;
         }
 
-        [HttpGet]
+
+        /// <summary>
+        /// Devuelve el listado completo de clientes activos.
+        /// </summary>
+        /// <returns></returns>
+        [HttpGet("ListadoClientes")]
+        [ProducesResponseType(typeof(IEnumerable<ClienteDto>), 200)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]        
         public async Task<IActionResult> Get()
         {
             var listaClientes = await _clienteRepository.GetClientesAsync();
@@ -32,8 +42,17 @@ namespace LibreriaElPortal_WebAPI.Controllers
             
             return Ok(listaClientes);
         }
-        
+
+        /// <summary>
+        /// Devuelve el cliente que coincida con el id proporcionado.
+        /// </summary>
+        /// <param name="id"></param>
+        /// <returns></returns>
         [HttpGet("{id}")]
+        [ProducesResponseType(typeof(ClienteDto), 200)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<IActionResult> GetCliente(int id)
         {
             var cliente = await _clienteRepository.GetClienteAsync(id);
@@ -46,7 +65,15 @@ namespace LibreriaElPortal_WebAPI.Controllers
             return Ok(cliente);
         }
 
+        /// <summary>
+        /// Recibe información de un cliente desde el cuerpo de la solicitud (ver ejemplo de "Request Body") para crear un nuevo registro.
+        /// </summary>
+        /// <param name="cliente"></param>
+        /// <returns></returns>
         [HttpPost]
+        [ProducesResponseType(StatusCodes.Status201Created, Type = typeof(ClienteDto))]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<IActionResult> CrearCliente([FromBody] AgregarClienteDto cliente)
         {
             var existe = await _clienteRepository.ExisteClienteByEmailAsync(cliente.Email);
@@ -63,7 +90,17 @@ namespace LibreriaElPortal_WebAPI.Controllers
             return CreatedAtAction(nameof(GetCliente), new { id = nuevoCliente.ClienteId }, nuevoCliente);
         }
 
+
+        /// <summary>
+        /// Recibe información de un cliente desde el cuerpo de la solicitud (ver ejemplo de "Request Body") para actualizar un registro existente.
+        /// </summary>
+        /// <param name="cliente"></param>
+        /// <returns></returns>
         [HttpPut]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+      
         public async Task<IActionResult> UpdateCliente([FromBody] ClienteDto cliente)
         {
             var extiste = await _clienteRepository.ExisteClienteAsync(cliente.ClienteId);
@@ -80,14 +117,27 @@ namespace LibreriaElPortal_WebAPI.Controllers
             return NoContent();
         }
 
+        /// <summary>
+        /// Recibe el id de un cliente por url y elimina el registro de la base de datos.
+        /// </summary>
+        /// <param name="clienteId"></param>
+        /// <returns></returns>
         [HttpDelete]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<IActionResult> DeleteCliente(int clienteId)
         {
             var extiste = await _clienteRepository.ExisteClienteAsync(clienteId);
-
             if (!extiste)
             {
-                return NotFound("No se encontró el cliente que intenta eliminar.");
+                return BadRequest("El cliente que intenta eliminar no existe.");
+            }
+
+            var ventasAsociadas = await _ventaRepository.GetVentasByClienteAsync(clienteId);
+            if (ventasAsociadas != null && ventasAsociadas.Count() > 0)
+            {
+                return BadRequest($"No se puede eliminar el cliente con ID: \"{clienteId}\" porque tiene ventas asociadas.");
             }
 
             var clienteEliminado = await _clienteRepository.DeleteClienteAsync(clienteId);
